+++ CHANGELOG.md
@@ -6,10 +6,10 @@
 
 ## 3.0.0 January 12, 2023
 
-- Fix #103: `ArrayableTrait::toArray()` returned an invalid result when no fields were specified (@ganigeorgiev)
 - Enh #115: Raise required PHP version to `^8.0`, move union type hints from annotations
   to methods' signatures (@xepozz, @vjik)
 - Enh #122: Add getters' support (keys like "getMyProperty()") to `ArrayHelper` (@vjik)
+- Bug #103: `ArrayableTrait::toArray()` returned an invalid result when no fields were specified (@ganigeorgiev)
 
 ## 2.1.0 August 20, 2022